dom/animation/test/chrome/test_running_on_compositor.html

<!doctype html>
<head>
<meta charset=utf-8>
<title>Bug 1045994 - Add a chrome-only property to inspect if an animation is
       running on the compositor or not</title>
<script type="application/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script type="application/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
<script type="application/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" type="text/css"
  href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
<style>
@keyframes anim {
  to { transform: translate(100px) }
}
div {
  /* Element needs geometry to be eligible for layerization */
  width: 100px;
  height: 100px;
  background-color: white;
}
</style>
</head>
<body>
<a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1045994"
  target="_blank">Mozilla Bug 1045994</a>
<script>
'use strict';

/** Test for bug 1045994 - Add a chrome-only property to inspect if an
    animation is running on the compositor or not **/

function addDiv(attrs) {
  var div = document.createElement('div');
  if (attrs) {
    for (var attrName in attrs) {
      div.setAttribute(attrName, attrs[attrName]);
    }
  }
  document.body.appendChild(div);
  return div;
}

const OMTAPrefKey = 'layers.offmainthreadcomposition.async-animations';
var omtaEnabled = SpecialPowers.DOMWindowUtils.layerManagerRemote &&
                  SpecialPowers.getBoolPref(OMTAPrefKey);

add_task(function* play_and_pause_from_style() {
  // FIXME: When we implement Element.animate, use that here instead of CSS
  // so that we remove any dependency on the CSS mapping.
  var div = addDiv({ style: 'animation: anim 100s' });
  var animation = div.getAnimations()[0];

  yield animation.ready;

  is(animation.isRunningOnCompositor, omtaEnabled,
     'Animation reports that it is running on the compositor'
     + ' during playback');

  div.style.animationPlayState = 'paused';

  yield animation.ready;

  is(animation.isRunningOnCompositor, false,
     'Animation reports that it is NOT running on the compositor'
     + ' when paused');
  div.parentNode.removeChild(div);
});
</script>
</body>
